+++ changelog.md
@@ -1,3 +1,11 @@
+# 1.5.5
+
+- changed Backblaze B2 service client dependency to a [fork](https://www.npmjs.com/package/@nichoth/backblaze-b2) that fixes the long standing Axios security issue &rarr; all credits: [nichoth](https://github.com/nichoth)
+
+# 1.5.4
+
+- added `getFileAsURL` to Amazon S3 adapter &rarr; all credits: [DennisSnijder](https://github.com/DennisSnijder)
+
 # 1.5.3
 
 - fix for issue #48 &rarr; all credits: [Pezmc](https://github.com/Pezmc)